Restructure components
Ohrwurm components have been restructered.
