+++ src/components/pages/index.ts
@@ -1,13 +1,14 @@
 //#region > Imports
 //> Pages
 // Import all components to export them for easy access from parent components
-import BasePage from "./BasePage";
+import OhrwurmPage from "./ohrwurm/index";
+import OhrwurmMemberPage from "./ohrwurm/MemberPage";
 import LoginPage from "./LoginPage";
 //#endregion
 
 //#region > Exports
 //> Pages
-export { BasePage, LoginPage };
+export { OhrwurmPage, OhrwurmMemberPage, LoginPage };
 //#endregion
 
 /**
