+++ src/components/pages/ohrwurm/MemberPage/index.tsx
@@ -18,16 +18,16 @@ import {
 } from "mdbreact";
 
 //> Store Types
-import { RootState } from "../../../store/reducers/index";
-import { UserState } from "../../../store/types";
+import { RootState } from "../../../../store/reducers/index";
+import { UserState } from "../../../../store/types";
 //> Store Actions
-import { loginAction } from "../../../store/actions/userActions";
+import { loginAction } from "../../../../store/actions/userActions";
 //> Service Registry
-import Services from "../../../serviceRegistry.json";
+import Services from "../../../../serviceRegistry.json";
 //> Components
-import { Breadcrumbs } from "../../atoms";
-import { LoginForm, ServiceGroup, MemberTable } from "../../molecules";
-import { MemberModal } from "../../organisms/modals";
+import { Breadcrumbs } from "../../../atoms";
+import { LoginForm, ServiceGroup, MemberTable } from "../../../molecules";
+import { MemberModal } from "../../../organisms/modals";
 //#endregion
 
 //#region > Interfaces
